+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+# 0.6.0-alpha
+* **BREAKING**: Period/length uses now seconds instead of minutes
+* Support of --debug flag which outputs some dev debug informations
+* Support of metrics who are not included in tags api (e.g. autoscaling metrics)
+`` `
+static:
+  - namespace: AWS/AutoScaling
+    region: eu-west-1
+    dimensions:
+     - name: AutoScalingGroupName
+       value: Test
+    metrics:
+      - name: GroupInServiceInstances
+        statistics:
+        - 'Minimum'
+        period: 60
+        length: 300
+```
+* **Breaking**: Config file uses new syntax to support static
+
+Before
+```yaml
+jobs:
+  - discovery:
+      region: eu-west-1
+    metrics:
+```
+After:
+```yaml
+discovery:
+  - region: eu-west-1
+    metrics:
+```
+
 # 0.5.0
 * Support of EFS - Elastic File System
 * Support of EBS - Elastic Block Storage